python-backend/FastAPI Server/requirements.txt
# 核心依赖
fastapi==0.112.0
uvicorn[standard]==0.31.0
python-multipart==0.0.9
numpy==2.0.1
pillow==11.3.0
opencv-python==4.12.0.88
onnxruntime==1.22.1

# 可选：如果模型需要PyTorch
torch==2.7.1+cu118
torchvision==0.22.1+cu118

# 其他工具
loguru==0.7.3  # 更好的日志
python-dotenv==1.0.1  # 环境变量管理